script: wait to delete logs until after test_results called
The CollectTestResults class will be reused to dump XML instead of
JSON, but to do so we first need to make sure that the log files
generated by each test are still available when CollectTestResults.log
is called. Otherwise, we won't be able to get all of the info we need
to populate the XML dumpfile.